modify file structure
