+++ scr/configs/Conditional_img_synthesis/contra_dcgan_cifar32_hinge_no.json
