+++ scr/configs/Conditional_img_synthesis/proj_dcgan_cifar32_hinge_no.json
