+++ scr/configs/Imagenet_experiments/contra_biggan_imagenet128_hinge_no.json
