+++ scr/configs/Imagenet_experiments/proj_biggan_imagenet128_hinge_no.json
