+++ scr/configs/Table1/ac_biggan_cifar32_hinge_no.json
