+++ scr/configs/Table1/ac_biggan_tiny64_hinge_no.json
