+++ scr/configs/Table1/contra_biggan_cifar32_hinge_no.json
