+++ scr/configs/Table1/contra_biggan_tiny64_hinge_no.json
