+++ scr/configs/Table1/proj_biggan_cifar32_hinge_no.json
