+++ scr/configs/Table1/proj_biggan_tiny64_hinge_no.json
