+++ scr/configs/Table2/contra_biggan_tiny64_hinge_no.json
