+++ scr/configs/Table2/proj_biggan_tiny64_hinge_no.json
