+++ scr/configs/Table2/proj_sngan_tiny64_hinge_no.json
