+++ scr/configs/Unconditional_img_synthesis/no_dcgan_cifar32_hinge_no.json
