+++ scr/models/biggan_deep.py
@@ -5,7 +5,7 @@
 # models/biggan_deep.py
 
 
-from models.model_ops import *
+from utils.model_ops import *
 
 import torch
 import torch.nn as nn
